backend/requirements.txt
# Backend Requirements for Counterfeit Product Detection API
# Install with: pip install -r backend/requirements.txt
# 
# NOTE: Python 3.13 compatibility
# - TensorFlow may not have official Python 3.13 support yet
# - If you encounter TensorFlow issues, consider using Python 3.11 or 3.12
# - Or install tensorflow-cpu instead

# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # For file uploads

# Machine Learning
# TensorFlow - Note: May not support Python 3.13 yet
# If you get errors, try: pip install tensorflow-cpu
# Or use Python 3.11/3.12 instead
tensorflow>=2.15.0  # Deep learning framework for image models

# HuggingFace transformers for text models
transformers>=4.35.2

# PyTorch (required by transformers) - Updated for Python 3.13
torch>=2.6.0  # Compatible with Python 3.13
torchvision>=0.21.0  # Updated to match torch version

# Data Processing
numpy>=1.24.3
pandas>=2.1.3
Pillow>=10.1.0  # Image processing

# Utilities
pydantic>=2.5.0  # Data validation (used by FastAPI)
python-dotenv>=1.0.0  # Environment variables

# Optional: Database (uncomment if using MongoDB)
# pymongo>=4.6.0
# motor>=3.3.2  # Async MongoDB driver

# Optional: Async task queue (uncomment if using Celery)
# celery>=5.3.4
# redis>=5.0.1

# Development
pytest>=7.4.3  # Testing framework
black>=23.11.0  # Code formatter
flake8>=6.1.0  # Linter

